Updated code according to @karusell review

+++ core/src/main/java/com/graphhopper/routing/util/spatialrules/SpatialRule.java
@@ -29,7 +29,7 @@ import java.util.List;
  */
 public interface SpatialRule {
 
-    int getMaxSpeed(ReaderWay readerWay, String transportationMode);
+    double getMaxSpeed(ReaderWay readerWay, String transportationMode);
 
     AccessValue isAccessible(ReaderWay readerWay, String transportationMode);
